rename common to commonRedux
