+++ src/js/bike-racks/actions/bikeRacksActions.js
@@ -1,6 +1,6 @@
 import BikeRack from '../records/BikeRack';
 import { name, path, indexFn } from '../constants/bikeRacksReduxConfig';
-import { all } from 'common/all';
+import { all } from 'commonRedux/all';
 
 const { actions } = all({ name, path, record: BikeRack, indexFn });
