+++ src/js/commonRedux/destroy.js
